required ROOT version now >=6.24, cmake >=3.9 (as for ROOT 6.24)
this is due to recent developments in KVTemplateEvent (allowing to use
KVParticle objects) and the consequent KVTemplateParticleCondition.
with ROOT < 6.24 not compiled in C++14 standard, autoloading of headers
fails when starting the kaliveda command-line.

it possibly could be resolved for ROOT<6.24 if compilation with C++14 is
forced (note that it says everywhere on the ROOT website that applications
based on ROOT MUST be compiled with the same C++ standard as ROOT -
whereas we now force C++14 compilation for KaliVeda, breaking this rule)